trigger/requirements.txt
fastapi
uvicorn
google-cloud-run
